+++ x-pack/plugins/secops/public/containers/errors/index.tsx
@@ -8,7 +8,7 @@ import { onError } from 'apollo-link-error';
 import uuid from 'uuid';
 
 import { store } from '../../store';
-import { addError } from '../../store/local/errors/actions';
+import { addError } from '../../store/local/app/actions';
 import * as i18n from './translations';
 
 export const errorLink = onError(({ graphQLErrors, networkError }) => {
