+++ x-pack/plugins/secops/public/mock/global_state.ts
@@ -12,6 +12,10 @@ export const mockGlobalState: State = {
   local: {
     app: {
       notesById: {},
+      errors: [
+        { id: 'error-id-1', title: 'title-1', message: 'error-message-1' },
+        { id: 'error-id-2', title: 'title-2', message: 'error-message-2' },
+      ],
     },
     hosts: {
       page: {
@@ -99,9 +103,5 @@ export const mockGlobalState: State = {
         },
       },
     },
-    errors: [
-      { id: 'error-id-1', title: 'title-1', message: 'error-message-1' },
-      { id: 'error-id-2', title: 'title-2', message: 'error-message-2' },
-    ],
   },
 };
