+++ x-pack/plugins/secops/public/store/local/app/actions.ts
@@ -11,3 +11,7 @@ import { Note } from '../../../lib/note';
 const actionCreator = actionCreatorFactory('x-pack/secops/local/app');
 
 export const updateNote = actionCreator<{ note: Note }>('UPDATE_NOTE');
+
+export const addError = actionCreator<{ id: string; title: string; message: string }>('ADD_ERRORS');
+
+export const removeError = actionCreator<{ id: string }>('REMOVE_ERRORS');
