--- x-pack/plugins/secops/public/store/local/errors/index.ts
@@ -1,15 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License;
- * you may not use this file except in compliance with the Elastic License.
- */
-
-import * as errorsActions from './actions';
-import * as errorsModel from './model';
-import * as errorsSelectors from './selectors';
-
-export { errorsActions };
-export { errorsModel };
-export { errorsSelectors };
-
-export * from './reducer';
