--- x-pack/plugins/secops/public/store/local/errors/selectors.ts
@@ -1,18 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License;
- * you may not use this file except in compliance with the Elastic License.
- */
-
-import { createSelector } from 'reselect';
-
-import { State } from '../../reducer';
-import { ErrorModel } from './model';
-
-const getErrors = (state: State): ErrorModel => state.local.errors;
-
-export const errorsSelector = () =>
-  createSelector(
-    getErrors,
-    errors => ({ errors })
-  );
